+++ httprunner/cli.py
@@ -91,14 +91,12 @@ def main_hrun():
 def main_locust():
     """ Performance test with locust: parse command line options and run commands.
     """
-    # monkey patch ssl at beginning to avoid RecursionError when running locust.
-    from gevent import monkey; monkey.patch_ssl()
-
-    import multiprocessing
-    import sys
-    from httprunner import logger
-
     try:
+        # monkey patch ssl at beginning to avoid RecursionError when running locust.
+        from gevent import monkey; monkey.patch_ssl()
+        import multiprocessing
+        import sys
+        from httprunner import logger
         from httprunner import locusts
     except ImportError:
         msg = "Locust is not installed, install first and try again.\n"