+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - add demo api/testcase/testsuite to new created scaffold project
 - update default `.gitignore` of new created scaffold project
+- add demo content to `debugtalk.py`/`.env` of new created scaffold project
 
 **Bugfixes**