+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - fix extend with testcase reference in format version 2
 - fix ImportError when locustio is not installed
+- fix YAMLLoadWarning by specify yaml loader
 
 ## 2.2.0 (2019-06-24)